Add simpler way to implement custom message types
All message implementations have to define and correctly initialize an
envelope field. This is not only prone to errors, but also dependent on
the internals of Flusswerk. An abstract base class encapsulates now the
technical specifics so that all implementations of the Message interface
now can focus on domain specific logic alone.