Add metrics collection to Spring Boot starter
The new Metrics bean supports the common collection tasks for total
processed messages and total execution time, each counted seperately for
successful processing, processing with errors and retry, processing with
errors that lead to abort processing. This bean is only created if the
user does not provide a Metrics bean first.