Release availableWorkers semaphore when message deserialization fails.
Previously every mesasge that failed to deserialize would decrease the
number of available workers by one. This led to a sitaution, where after
a few messages (as many as there are available workers configured) the
engine would just refuse to accept any incoming messages, since it would
wait for eternity for the `availableWorkers` to become available again.